+++ src/Coil/CUDAKernels/MInductance.cu
@@ -13,7 +13,7 @@ void CalculateMutualInductanceConfigurations(long long configCount, long long po
                                              const CoilPairPositionData *configArr, TYPE *inductanceArr)
 {
     unsigned int index = threadIdx.x;
-    long long global_index = blockIdx.x * blockDim.x + index;
+    long long global_index = (long long) blockIdx.x * blockDim.x + index;
 
     if(global_index >= configCount * pointCount)
         return;
@@ -191,7 +191,7 @@ void Calculate_mutual_inductance_configurations(long long configCount, long long
         recordStartPoint();
     #endif
 
-    int blocks = ceil(double(configCount * pointCount) / NTHREADS);
+    int blocks = int(std::ceil(double((long long) configCount * pointCount) / NTHREADS));
 
     getBuffers(configCount);
