+++ src/CoilGroup/CUDAKernels/FieldGGroup.cu
@@ -14,7 +14,7 @@ void calculateGradientGroup(long long opCount, long long coilIndex,
                             MatrixData *resArr)
 {
     unsigned int index = threadIdx.x;
-    long long globalIndex = blockIdx.x * blockDim.x + index;
+    long long globalIndex = (long long) blockIdx.x * blockDim.x + index;
 
     if(globalIndex >= opCount)
         return;
@@ -219,7 +219,7 @@ void Calculate_hardware_accelerated_g_group(long long coilCount, long long opCou
         recordStartPoint();
     #endif
 
-    long long blocks = ceil(double(opCount) / NTHREADS);
+    int blocks = int(std::ceil(double(opCount) / NTHREADS));
 
     getBuffers(coilCount, opCount);
